fix minus operator being applied to an unsigned type
- still treated as unsigned regardless so we cast it to a signed type

+++ xlive/Blam/Engine/interface/screens/screen_cartographer_account_manager.cpp
@@ -158,7 +158,7 @@ c_cartographer_account_manager_edit_list::c_cartographer_account_manager_edit_li
 
 	// no fucking clue what's this, maybe related to this->m_slot_2.c_slot_vtbl data offset
 	// because this->m_slot_2.c_slot_vtbl is at offset 708
-	if ((void*)this == (void*)-offsetof(c_cartographer_account_manager_edit_list, m_slot_2)) {
+	if ((void*)this == (void*)-(int32)offsetof(c_cartographer_account_manager_edit_list, m_slot_2)) {
 		linker_type2.link(nullptr);
 	}
 	else {
